+++ src/layouts/Main.jade
@@ -47,6 +47,13 @@ html(lang="en")
                             .list-row__content
                                 span.display-block Typearea
                                 span.display-block.fs-body-1.tc-black-2 แก้ไข Typearea
+                        li.list-row.list-row--has-separator
+                            .list-row__primary
+                                a(href="#/chronic")
+                                    i.icon.icon--l.icon--red.icon--circled.mdi.mdi-needle
+                            .list-row__content
+                                span.display-block Chronic
+                                span.display-block.fs-body-1.tc-black-2 แก้ไข Chronic
 
             div(flex-item="9")
                 .card
