+++ src/pages/Index.jade
@@ -8,3 +8,5 @@ block scripts
     script(src="../main/IndexService.js")
     script(src="../typearea/TypeareaController.js")
     script(src="../typearea/TypeareaService.js")
+    script(src="../chronic/ChronicController.js")
+    script(src="../chronic/ChronicService.js")
